Refactor CLI and manifest handling for clearer modularity
