+++ pcdset/cli/__init__.py
@@ -1,0 +1,6 @@
+"""Command line interface for :mod:`pcdset`."""
+from __future__ import annotations
+
+from .app import app, create_app
+
+__all__ = ["app", "create_app"]
