+++ pcdset/manifest/__init__.py
@@ -1,0 +1,19 @@
+"""Utilities for reading and writing dataset manifests."""
+from __future__ import annotations
+
+from .builders import assign_splits, build_simple_entries
+from .examples import build_example_manifest, build_example_manifest_shapenet
+from .io import write_manifest
+from .loaders import load_entries, load_entries_shapenet
+from .models import Entry
+
+__all__ = [
+    "Entry",
+    "assign_splits",
+    "build_simple_entries",
+    "build_example_manifest",
+    "build_example_manifest_shapenet",
+    "load_entries",
+    "load_entries_shapenet",
+    "write_manifest",
+]
