+++ pcdset/profiles/__init__.py
@@ -1,8 +1,15 @@
 """Profile implementations for :mod:`pcdset`."""
 from __future__ import annotations
 
-__all__ = ["PCNProfile", "ShapeNetProfile", "BaseProfile"]
+__all__ = [
+    "BaseProfile",
+    "PCNProfile",
+    "ShapeNetProfile",
+    "get_profile_class",
+    "iter_profiles",
+]
 
 from .base import BaseProfile
 from .pcn import PCNProfile
+from .registry import get_profile_class, iter_profiles
 from .shapenet import ShapeNetProfile
