+++ pcdset/profiles/shapenet.py
@@ -23,7 +23,7 @@ from ..ops import (
     dedup as op_dedup,
 )
 from ..utils.logging import logger
-from ..utils.manifest import Entry
+from ..manifest import Entry
 from ..utils import taxonomy
 from .base import BaseProfile
